This is the 4th version of the project
*The unstable behavior is possibly related to the serializers.ChoiceField attribute.

If you look at the repository,
https://github.com/serg-lubeshko/Lubeshko_t_4.git.
https://github.com/serg-lubeshko/newLubeshko_t4.git

Redid the project 4 times. In old project the problem there was that when
deleting the migration and the database,
there was an error with the creating of a new database.

first, I move all models to a new folder
now, I have a different name for the application